+++ src/pages/workflows/Workflows.tsx
@@ -1,9 +1,9 @@
 import DashboardLayout from '../../components/layouts/DashboardLayout';
 import { Workflow } from './models/Workflow';
-import WorkflowUnit from './WorkflowUnit';
-import { WorkflowCommandBar } from './WorkflowCommandBar';
+import WorkflowUnit from './components/workflow-unit/WorkflowUnit';
+import { WorkflowCommandBar } from './components/workflow-command-bar/WorkflowCommandBar';
 import * as icons from '@ant-design/icons/lib/icons';
-import { ScheduledWorkflowBar } from './ScheduledWorkflowBar';
+import { ScheduledWorkflowBar } from './components/scheduled-workflow-bar/ScheduledWorkflowBar';
 
 function Workflows() {
     return (
