+++ src/pages/workflows/components/workflow-unit/WorkflowUnitProps.ts
@@ -1,4 +1,4 @@
-import { Workflow } from "./models/Workflow";
+import { Workflow } from "../../models/Workflow";
 
 export interface WorkflowUnitProps {
     workflow: Workflow,
